version 0.1.6 1)change method __add__ in class Product 2)wrote additional tests 3)add annotations in class_category_and_iter and class_product_and_descendants 4)remove .coverage
