+++ src/class_product_and_descendants.py
@@ -1,52 +1,7 @@
-class Category:
-    """Класс категории товаров"""
-    number_of_categories = 0
-
-    name: str
-    description: str
-    products: list
-
-    def __init__(self, name, description, products):
-        self.name = name
-        self.description = description
-        self.__products = products
-
-        Category.number_of_categories += 1
-
-    @property
-    def goods(self):
-        list_products = []
-        for product in self.__products:
-            list_products.append(f"{product.name}, {int(product.price)} руб. Остаток: {product.quantity} шт.")
-        return list_products
-
-    @property
-    def products(self):
-        return self.__products
-
-    def add_products(self, product):
-        if isinstance(product, Product):
-            self.__products.append(product)
-
-    def __len__(self):
-        return len(self.__products)
-
-    def __repr__(self):
-        return f"Category({self.name}; {self.description}; {self.products})"
-
-    def __str__(self):
-        return f"{self.name}, количество продуктов: {len(self)} шт."
-
-
 class Product:
     """Класс товара"""
-    name: str
-    description: str
-    price: float
-    quantity: int
-    color: str
 
-    def __init__(self, name, description, price, quantity, color):
+    def __init__(self, name: str, description: str, price: float, quantity: int, color: str):
         self.name = name
         self.description = description
         self.__price = price
@@ -88,7 +43,7 @@ class Product:
                     print("Пожалуйста, введите корректный ответ: y или n(y-да, n-нет)")
 
     def __add__(self, other):
-        if isinstance(other, self.__class__):
+        if type(other) == type(self):
             return self.price * self.quantity + other.price * other.quantity
         raise TypeError
 
@@ -100,7 +55,8 @@ class Product:
 
 
 class Smartphone(Product):
-    def __init__(self, name, description, price, quantity, color, performance, model, memory):
+    def __init__(self, name: str, description: str, price: float, quantity: int,
+                 color: str, performance: str, model: str, memory: str):
         super().__init__(name, description, price, quantity, color)
         self.performance = performance
         self.model = model
@@ -108,23 +64,8 @@ class Smartphone(Product):
 
 
 class LawnGrass(Product):
-    def __init__(self, name, description, price, quantity, color, country, germination_period):
+    def __init__(self, name: str, description: str, price: float, quantity: int,
+                 color: str, country: str, germination_period: str):
         super().__init__(name, description, price, quantity, color)
         self.country = country
         self.germination_period = germination_period
-
-
-class IterationProducts:
-    def __init__(self, category):
-        self.category = category["products"]
-
-    def __iter__(self):
-        self.current_value = -1
-        return self
-
-    def __next__(self):
-        if self.current_value + 1 < len(self.category):
-            self.current_value += 1
-            return self.category[self.current_value]
-        else:
-            raise StopIteration
